requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
certifi==2025.7.14
charset-normalizer==3.4.2
django==4.2.21
djangorestframework==3.15.2
exceptiongroup==1.3.0
factory-boy==3.3.3
Faker==35.2.2
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
packaging==25.0
pluggy==1.5.0
pytest==8.3.5
pytest-django==4.11.1
pytest-factoryboy==2.7.0
python-dateutil==2.9.0.post0
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
tomli==2.2.1
typing-extensions==4.13.2
urllib3==2.2.3
